style: rename model modules
