+++ project/app/api/crud.py
@@ -1,7 +1,7 @@
 from typing import Dict, List, Union
 
-from app.models.pydantic import SummaryPayloadSchema, SummaryUpdatePayloadSchema
-from app.models.tortoise import TextSummary
+from app.models.pydantic_model import SummaryPayloadSchema, SummaryUpdatePayloadSchema
+from app.models.tortoise_model import TextSummary
 
 
 async def post(payload: SummaryPayloadSchema) -> int:
